+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Let's take the `x.β.α` version number:
 
 ## α.β.x (coming soon)
 
+## 1.14.2 (2018-12-12)
+* add component `other-hours`
+* add component `pills`
+* component `text`: add social network data in text
+
 ## 1.14.1 (2018-12-05)
 
 * modules `rush-half-hours` : Fix bug when days starts at half hour